src/modules/client-adm/usecase/add-client/find-client/find-client.usecase.spec.ts
import Id from "../../../../@shared/domain/entity/valeu-object/id.value-object";
import Client from "../../../domain/client.entity";
import FindClientUseCase from "./find-client.usecase";

const client = new Client({
    id: new Id("1"),
    name: "Taynara",
    email: "x@anote.com.br",
    address: "Pandora, BorderLands"
});

const MockRepository = () => {
    return {
        add: jest.fn(),
        find: jest.fn().mockReturnValue(Promise.resolve(client))
    }
}

describe("Teste de caso de uso de localizar client", () => {
    it('Deve localizar um clienbt', async () => {
        const repository = MockRepository();
        const usecase = new FindClientUseCase(repository);

        const input = {
            id: "1"
        };

        const result = await usecase.execute(input);

        expect(repository.find).toHaveBeenCalled();
        expect(result.id).toBe(input.id);
        expect(result.name).toBe(client.name);
        expect(result.email).toBe(client.email);
        expect(result.address).toBe(client.address);
        expect(result.createdAt).toBe(client.createdAt);
        expect(result.updatedAt).toBe(client.updatedAt);
    })
});